Captures actual primary key column, fixes static find() method
This commit refactors ActiveRecord to use a new static property `primary_key_name` that allows subclassers to specify the name of the primary key column other than 'id'. The class also detects the actual primary key if disagreeing with the declared one, and stores it into a dictionary `actual_primary_key_names` with the class name as key. Finally, the static `find()` method is refactored to actually look for the actual class receiving the method call, rendering the second argument unnecessary (and deprecated).

Fixes #9.